Reorganizes directory structure. For scale.
